Split demo to computations and visualization (complete).
Make computations using run_analysis.m, which does not create any figures. Do visualization using view_analysis.m which does not do any intensive computations. The former requires a strong machine (especially for big data); the latter can be done anywhere, as long as the ram is big enough to load the data (-to-be-visualized).